+++ ICD.Connect.Audio/VolumePoints/IVolumePointSettings.cs
@@ -53,5 +53,10 @@ namespace ICD.Connect.Audio.VolumePoints
 		/// Determines what muting this volume point will do (mute audio output, mute microphones, etc).
 		/// </summary>
 		eMuteType MuteType { get; set; }
+
+		/// <summary>
+		/// Determines if the privacy mute control will be driven by the control system, and/or drive the control system.
+		/// </summary>
+		ePrivacyMuteFeedback PrivacyMuteMask { get; set; }
 	}
 }